CLI/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "aira-scanner"
version = "1.2.0"
description = "AIRA — AI-Induced Risk Audit: Static analysis for AI-generated code failure patterns"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "William M. Parris", email = "bill@bageltech.net"}
]
keywords = [
    "ai", "code-quality", "static-analysis", "governance",
    "ai-safety", "audit", "constitutional-ai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
aira = "aira.cli:main"

[project.urls]
Homepage = "https://bageltech.net"
Repository = "https://github.com/bageltech/aira-scanner"
"Bug Tracker" = "https://github.com/bageltech/aira-scanner/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["aira*"]
